+++ main/stats.c
@@ -1,6 +1,6 @@
 #include "stats.h"
 
-void stats(uint16_t * arr, unsigned int len, uint64_t * avg, uint16_t * min, uint16_t * max) {
+void stats(int32_t * arr, unsigned int len, int64_t * avg, int32_t * min, int32_t * max) {
     *min = arr[0];
     *max = arr[0];
     *avg = 0;
